Current APIs Working
The menu and both APIs are working. Some of the settings pages are in place. Will begin reincorporating the import functions, add more specific pricing functions for those who support it (outside of reseller dashboards), fix the "help" sections, and more coming soon. Much of this is already in place, needs to be added back to the main plugin file.